third_party/blink/web_tests/platform/generic/fast/frames/frameset-frameborder-boolean-values-expected.txt

Check frameset frameborder attribute allows 1,yes for true and 0,no for false

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

frameborder=1 is true
PASS frames[0].offsetWidth is 145
PASS frames[1].offsetWidth is 145


frameborder=yes is true
PASS frames[0].offsetWidth is 145
PASS frames[1].offsetWidth is 145


frameborder=0 is false
PASS frames[0].offsetWidth is 150
PASS frames[1].offsetWidth is 150


frameborder=no is false
PASS frames[0].offsetWidth is 150
PASS frames[1].offsetWidth is 150


Invalid values like frameborder=false are ignored
PASS frames[0].offsetWidth is 147
PASS frames[1].offsetWidth is 147


Invalid values like frameborder=2 are ignored
PASS frames[0].offsetWidth is 147
PASS frames[1].offsetWidth is 147


PASS successfullyParsed is true

TEST COMPLETE

     
